+++ apps/modernization-ui/src/apps/patient/file/demographics/race/patientRace.ts
@@ -1,5 +1,5 @@
 import { PatientFileService } from 'generated';
-import { RaceDemographic } from 'libs/patient/demographics/race/race';
+import { RaceDemographic } from 'libs/patient/demographics/race';
 
 const patientRace = (patient: number): Promise<Array<RaceDemographic>> =>
     PatientFileService.raceDemographics({ patient }).then();
