+++ apps/modernization-ui/src/libs/patient/demographics/race/index.ts
@@ -1,0 +1,4 @@
+export { RaceDemographicCard } from './RaceDemographicCard';
+export type { RaceDemographicCardProps } from './RaceDemographicCard';
+
+export type { RaceDemographic } from './race';
